# [Numpy] FFI for linalg.lstsq/qr